# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Say It Anyway API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Room {
  id: string;
  code: string;
  createdAt: string;
  isFixedRoom: boolean;
  activeMode: string;
  /** @nullable */
  activeLevel: number | null;
  activeRelationshipFilter: string;
  /** @nullable */
  currentCardId: string | null;
  currentCardIndex: number;
  shuffledDeckOrder: string[];
  afterDarkUnlocked: boolean;
  createdByPlayerName: string;
  updatedAt: string;
}

export interface RoomInput {
  /** Custom code (for fixed SAYIT room). If omitted, one is generated. */
  code?: string;
  createdByPlayerName: string;
  isFixedRoom?: boolean;
}

export interface RoomStateUpdate {
  activeMode?: string;
  /** @nullable */
  activeLevel?: number | null;
  activeRelationshipFilter?: string;
  /** @nullable */
  currentCardId?: string | null;
  currentCardIndex?: number;
  shuffledDeckOrder?: string[];
  afterDarkUnlocked?: boolean;
  updatedByPlayerName?: string;
}

export interface Player {
  id: string;
  roomId: string;
  displayName: string;
  joinedAt: string;
  lastSeenAt: string;
}

export interface PlayerInput {
  displayName: string;
  /** If provided, update existing player; otherwise create new */
  playerId?: string;
}

export interface RoomJoinResult {
  room: Room;
  player: Player;
}

export interface CardHistoryInput {
  cardId: string;
  mode: string;
  /** @nullable */
  level?: number | null;
  advancedByPlayerName: string;
}

export interface CardHistory {
  id: string;
  roomId: string;
  cardId: string;
  mode: string;
  /** @nullable */
  level?: number | null;
  shownAt: string;
  advancedByPlayerName: string;
}