# lib/api-zod/src/generated/types/cardHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Say It Anyway API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CardHistory {
  id: string;
  roomId: string;
  cardId: string;
  mode: string;
  /** @nullable */
  level?: number | null;
  shownAt: string;
  advancedByPlayerName: string;
}