# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Say It Anyway API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: string;
  roomId: string;
  displayName: string;
  joinedAt: string;
  lastSeenAt: string;
}