# lib/api-zod/src/generated/types/playerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Say It Anyway API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerInput {
  displayName: string;
  /** If provided, update existing player; otherwise create new */
  playerId?: string;
}