# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Say It Anyway API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: string;
  code: string;
  createdAt: string;
  isFixedRoom: boolean;
  activeMode: string;
  /** @nullable */
  activeLevel: number | null;
  activeRelationshipFilter: string;
  /** @nullable */
  currentCardId: string | null;
  currentCardIndex: number;
  shuffledDeckOrder: string[];
  afterDarkUnlocked: boolean;
  createdByPlayerName: string;
  updatedAt: string;
}