# lib/api-zod/src/generated/types/roomInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Say It Anyway API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoomInput {
  /** Custom code (for fixed SAYIT room). If omitted, one is generated. */
  code?: string;
  createdByPlayerName: string;
  isFixedRoom?: boolean;
}