# lib/api-zod/src/generated/types/roomJoinResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Say It Anyway API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';
import type { Room } from './room';

export interface RoomJoinResult {
  room: Room;
  player: Player;
}