# lib/api-zod/src/generated/types/roomStateUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Say It Anyway API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoomStateUpdate {
  activeMode?: string;
  /** @nullable */
  activeLevel?: number | null;
  activeRelationshipFilter?: string;
  /** @nullable */
  currentCardId?: string | null;
  currentCardIndex?: number;
  shuffledDeckOrder?: string[];
  afterDarkUnlocked?: boolean;
  updatedByPlayerName?: string;
}